+++ app/tests/components/Header.test.tsx
@@ -1,5 +1,5 @@
 import userEvent from "@testing-library/user-event";
-import { render, screen } from "tests/test-utils";
+import { render, screen } from "tests/react-utils";
 
 import Header from "src/components/Header";
